pyproject.toml
[tool.poetry]
name = "website"
version = "1.0.0"
description = ""
authors = ["BD103 <dont@stalk.me>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
libsass = "^0.20.1"
jinja2 = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
